requirements.txt
Django==4.2.7
rembg
Pillow==10.1.0
